Created all endpoints, added validation, made corresponding changes.

+++ bank-system/src/main/java/com/mybanksystem/transaction/model/TransactionContext.java
@@ -2,14 +2,16 @@ package com.mybanksystem.transaction.model;
 
 import com.mybanksystem.transaction.model.enumeration.TransactionType;
 
+import java.math.BigDecimal;
+
 public class TransactionContext {
     private TransactionType type;
     private Long fromId;
     private Long toId;
-    private Double amount;
+    private BigDecimal amount;
     private Long bankId;
 
-    public TransactionContext(TransactionType type, Long fromId, Long toId, Double amount, Long bankId) {
+    public TransactionContext(TransactionType type, Long fromId, Long toId, BigDecimal amount, Long bankId) {
         this.type = type;
         this.fromId = fromId;
         this.toId = toId;
@@ -29,7 +31,7 @@ public class TransactionContext {
         return toId;
     }
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
